+++ tket/tests/Circuit/test_CircPool.cpp
@@ -15,12 +15,13 @@
 #include <catch2/catch_test_macros.hpp>
 
 #include "Circuit/CircPool.hpp"
+#include "Predicates/Predicates.hpp"
 #include "Simulation/CircuitSimulator.hpp"
 
 namespace tket {
 namespace test_CircPool {
 
-SCENARIO("CircPool identities are correct") {
+SCENARIO("Simple CircPool identities") {
   Circuit orig, res;
 
   GIVEN("tk1_to_tk1") {
@@ -47,6 +48,130 @@ SCENARIO("CircPool identities are correct") {
   auto u_orig = tket_sim::get_unitary(orig);
   auto u_res = tket_sim::get_unitary(res);
   REQUIRE(u_res.isApprox(u_orig));
+}
+
+SCENARIO("TK2_using_normalised_TK2") {
+  Expr e1, e2, e3;
+  Sym asym = SymEngine::symbol("a");
+  Expr a(asym);
+  Sym bsym = SymEngine::symbol("b");
+  Expr b(bsym);
+  Sym csym = SymEngine::symbol("c");
+  Expr c(csym);
+
+  GIVEN("Normalised concrete angles (1)") {
+    e1 = .3;
+    e2 = .1;
+    e3 = .05;
+  }
+  GIVEN("Normalised concrete angles (2)") {
+    e1 = 0.32;
+    e2 = 0.31;
+    e3 = -0.3;
+  }
+  GIVEN("Not normalised concrete angles (1)") {
+    e1 = .3;
+    e2 = .4;
+    e3 = .45;
+  }
+  GIVEN("Not normalised concrete angles (2)") {
+    e1 = .3;
+    e2 = 1.4;
+    e3 = .489;
+  }
+  GIVEN("Not normalised concrete angles (3)") {
+    e1 = 2.3;
+    e2 = 3.4;
+    e3 = .489;
+  }
+  GIVEN("Not normalised concrete angles (4)") {
+    e1 = .3;
+    e2 = -.2;
+    e3 = .1;
+  }
+  GIVEN("Not normalised concrete angles (5)") {
+    e1 = -.3;
+    e2 = -.2;
+    e3 = .1;
+  }
+  GIVEN("Not normalised concrete angles (6)") {
+    e1 = .3;
+    e2 = .2;
+    e3 = -.3;
+  }
+  GIVEN("Not normalised concrete angles (7)") {
+    e1 = 0;
+    e2 = 0;
+    e3 = -1.2;
+  }
+  GIVEN("Not normalised concrete angles (8)") {
+    e1 = 0.1;
+    e2 = 0.3;
+    e3 = 0.2;
+  }
+  GIVEN("Symbolic angles (1)") {
+    e1 = a;
+    e2 = 3.4;
+    e3 = .489;
+  }
+  GIVEN("Symbolic angles (2)") {
+    e1 = a;
+    e2 = b;
+    e3 = 2.42;
+  }
+  GIVEN("Symbolic angles (3)") {
+    e1 = 2.3;
+    e2 = b;
+    e3 = 1.489;
+  }
+  GIVEN("Symbolic angles (4)") {
+    e1 = 2.3;
+    e2 = 123.08174;
+    e3 = c;
+  }
+  GIVEN("Symbolic angles (5)") {
+    e1 = a;
+    e2 = 123.08174;
+    e3 = c;
+  }
+  GIVEN("Symbolic angles (6)") {
+    e1 = 0.10012;
+    e2 = b;
+    e3 = c;
+  }
+
+  Circuit orig(2);
+  orig.add_op<unsigned>(OpType::TK2, {e1, e2, e3}, {0, 1});
+  Circuit res = CircPool::TK2_using_normalised_TK2(e1, e2, e3);
+
+  REQUIRE(NormalisedTK2Predicate().verify(res));
+
+  // check unitary identity
+  auto symset = orig.free_symbols();
+  std::vector<Sym> symbols(symset.begin(), symset.end());
+  if (symbols.empty()) {
+    auto u_orig = tket_sim::get_unitary(orig);
+    auto u_res = tket_sim::get_unitary(res);
+    REQUIRE(u_res.isApprox(u_orig));
+  } else {
+    std::vector<double> rands{0.1231, 2.3124, 34.23, 2.23, 3.15, 1.2, 0.93};
+    // substitute random values for symbolics and check equality
+    unsigned i = 0;
+    while (i + symbols.size() <= rands.size()) {
+      symbol_map_t symmap;
+      for (unsigned j = 0; j < symbols.size(); ++j) {
+        symmap[symbols[j]] = rands[i + j];
+      }
+      Circuit res_sub = res;
+      Circuit orig_sub = orig;
+      res_sub.symbol_substitution(symmap);
+      orig_sub.symbol_substitution(symmap);
+      auto u_orig = tket_sim::get_unitary(orig_sub);
+      auto u_res = tket_sim::get_unitary(res_sub);
+      REQUIRE(u_res.isApprox(u_orig));
+      ++i;
+    }
+  }
 }
 
 }  // namespace test_CircPool